More restructuring and some fixes
function that werent used or made since like create and delete or missing gets on customizable tables
