+++ code/jvm/src/main/kotlin/pt/arbitros/arbnet/repository/adaptable_repos/FunctionRepository.kt
@@ -1,6 +1,11 @@
 package pt.arbitros.arbnet.repository.adaptable_repos
 
+import pt.arbitros.arbnet.domain.universal.Function
+
 interface FunctionRepository {
+
+    fun getAllFunctions(): List<Function>
+
     fun getFunctionIdByName(roleName: String): Int?
 
     fun getFunctionNameById(roleId: Int): String
