+++ code/jvm/src/main/kotlin/pt/arbitros/arbnet/repository/jdbi/TransactionJdbi.kt
@@ -11,6 +11,11 @@ import pt.arbitros.arbnet.repository.adaptable_repos.EquipmentRepository
 import pt.arbitros.arbnet.repository.adaptable_repos.FunctionRepository
 import pt.arbitros.arbnet.repository.adaptable_repos.PositionRepository
 import pt.arbitros.arbnet.repository.adaptable_repos.RoleRepository
+import pt.arbitros.arbnet.repository.jdbi.adaptable_repos.CategoryRepositoryJdbi
+import pt.arbitros.arbnet.repository.jdbi.adaptable_repos.EquipmentRepositoryJdbi
+import pt.arbitros.arbnet.repository.jdbi.adaptable_repos.FunctionRepositoryJdbi
+import pt.arbitros.arbnet.repository.jdbi.adaptable_repos.PositionRepositoryJdbi
+import pt.arbitros.arbnet.repository.jdbi.adaptable_repos.RoleRepositoryJdbi
 
 class TransactionJdbi(
     private val handle: Handle,
