+++ code/jvm/src/main/kotlin/pt/arbitros/arbnet/repository/jdbi/adaptable_repos/FunctionRepositoryJdbi.kt
@@ -1,12 +1,17 @@
-package pt.arbitros.arbnet.repository.jdbi
+package pt.arbitros.arbnet.repository.jdbi.adaptable_repos
 
 import org.jdbi.v3.core.Handle
 import org.jdbi.v3.core.kotlin.mapTo
+import pt.arbitros.arbnet.domain.universal.Function
 import pt.arbitros.arbnet.repository.adaptable_repos.FunctionRepository
 
 class FunctionRepositoryJdbi(
     private val handle: Handle,
 ) : FunctionRepository {
+    override fun getAllFunctions(): List<Function> {
+        TODO("Not yet implemented")
+    }
+
     override fun getFunctionIdByName(functionName: String): Int? =
         handle
             .createQuery(
